+++ SourceCode/Data/MeetingParticipant.cs
@@ -20,8 +20,6 @@ public class MeetingParticipant
     public bool ParticipateDay4 { get; set; }
     public bool ParticipateDay5 { get; set; }
     public bool ParticipateDay6 { get; set; }
-    [Obsolete("Use LatestArrivalTime")]
-    public DateTime? ArrivalTime { get; set; }
     public TimeOnly? LatestArrivalTime { get; set; }
     public TimeOnly? EarliestDepartureTime { get; set; }
 
@@ -90,6 +88,34 @@ public static class MeetingParticipantExtensions
 
     public static string NameWithCity(this MeetingParticipant participant) =>
         participant.Person is not null ? participant.Person.NameWithCity() : string.Empty;
+
+    public static MeetingParticipant DefaultParticipant(this Meeting meeting, int personId)
+    {
+        var participant = new MeetingParticipant
+        {
+            MeetingId = meeting.Id,
+            PersonId = personId,
+            LatestArrivalTime = meeting.LatestArrivalTimeWithModules,
+            EarliestDepartureTime = meeting.EarliestDepartureTimeWithModules,
+        };
+        participant.SetDefaultParticipationDays(meeting);
+        if (meeting.Layouts.Count == 1)
+        {
+            participant.LayoutParticipations.Add(new LayoutParticipant() { LayoutId = meeting.Layouts.First().Id, PersonId = personId });
+        }
+        return participant;
+    }
+
+    private static void SetDefaultParticipationDays(this MeetingParticipant participant, Meeting meeting)
+    {
+        var days = meeting.DurationDays();
+        if (days > 0) participant.ParticipateDay1 = true;
+        if (days > 1) participant.ParticipateDay2 = true;
+        if (days > 2) participant.ParticipateDay3 = true;
+        if (days > 3) participant.ParticipateDay4 = true;
+        if (days > 4) participant.ParticipateDay5 = true;
+        if (days > 5) participant.ParticipateDay6 = true;
+    }
 
 }